@@ -3,7 +3,7 @@ name = "peridot"
 description = "Peridot Engine"
 version = "0.1.0"
 authors = ["S.Percentage <Syn.Tri.Naga@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [features]
 debug = ["bedrock/VK_EXT_debug_utils"]
